U4SM Interface
The U4SM Interface provides an entry point to access U4SM Web Interfaces and also an interface to access the U4SM API, allowing you to update and add records via the tool instead of via the web interface. This tool is intended for use by U4SM Administrators.